+++ sdk/azurestack/arm-azurestack/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 3.0.0-beta.4 (Unreleased)
+## 3.0.0-beta.5 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,12 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 3.0.0-beta.4 (2022-10-08)
+
+**Bugs Fixed**
+
+  -  revert credential scopes
 
 ## 3.0.0-beta.3 (2022-09-30)