+++ sdk/servicebus/service-bus/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 7.7.3 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 7.7.2 (2022-10-11)
 
 ### Other Changes